This fixes a few potential overflows and other issues that were found by MSVC warnings.

Related PR: https://github.com/ginkgo-project/ginkgo/pull/666

+++ include/ginkgo/core/matrix/hybrid.hpp
@@ -245,10 +245,10 @@ public:
          * @param percent  the row_nnz[floor(num_rows*percent)] is the number of
          *                 stored elements per row of the ell part
          */
-        explicit imbalance_limit(float percent = 0.8) : percent_(percent)
+        explicit imbalance_limit(double percent = 0.8) : percent_(percent)
         {
-            percent_ = std::min(percent_, 1.0f);
-            percent_ = std::max(percent_, 0.0f);
+            percent_ = std::min(percent_, 1.0);
+            percent_ = std::max(percent_, 0.0);
         }
 
         size_type compute_ell_num_stored_elements_per_row(
@@ -276,7 +276,7 @@ public:
         auto get_percentage() const { return percent_; }
 
     private:
-        float percent_;
+        double percent_;
     };
 
     /**
@@ -289,7 +289,7 @@ public:
         /**
          * Creates a imbalance_bounded_limit strategy.
          */
-        imbalance_bounded_limit(float percent = 0.8, float ratio = 0.0001)
+        imbalance_bounded_limit(double percent = 0.8, double ratio = 0.0001)
             : strategy_(imbalance_limit(percent)), ratio_(ratio)
         {}
 
@@ -319,7 +319,7 @@ public:
 
     private:
         imbalance_limit strategy_;
-        float ratio_;
+        double ratio_;
     };
